clean up method and component signatures, and move updateStates into singleFormField utils as updateFormFieldStates
